requirements.txt
asgiref==3.10.0
beautifulsoup4==4.12.3
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.0.14
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.5.1
fake-useragent==1.5.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.11
lxml==5.3.0
numpy==2.2.6
packaging==25.0
pandas==2.2.2
pillow==12.0.0
playwright==1.45.0
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
pyee==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
